Video Frame Extraction/requirements.txt
Flask==2.3.2
opencv-python==4.8.0.76
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0